+++ contracts/MMInterface.sol
@@ -7,10 +7,10 @@ contract MMInterface is Instantiator
 {
   enum state { WaitingProofs, WaitingReplay, FinishedReplay }
 
-  function newHash(uint32 _index) public view returns (bytes32);
+  function newHash(uint256 _index) public view returns (bytes32);
   function instantiate(address _provider, address _client,
-                       bytes32 _initialHash) public returns (uint32);
-  function stateIsWaitingProofs(uint32 _index) public view returns(bool);
-  function stateIsWaitingReplay(uint32 _index) public view returns(bool);
-  function stateIsFinishedReplay(uint32 _index) public view returns(bool);
+                       bytes32 _initialHash) public returns (uint256);
+  function stateIsWaitingProofs(uint256 _index) public view returns(bool);
+  function stateIsWaitingReplay(uint256 _index) public view returns(bool);
+  function stateIsFinishedReplay(uint256 _index) public view returns(bool);
 }
