+++ contracts/MachineInterface.sol
@@ -5,7 +5,7 @@ contract MachineInterface
 {
   event StepGiven(uint8 exitCode);
 
-  function endStep(address, uint32, uint8) internal returns (uint8);
+  function endStep(address, uint256, uint8) internal returns (uint8);
 
-  function step(address, uint32) public returns (uint8);
+  function step(address, uint256) public returns (uint8);
 }
